Backport PR #8196: DEV: Allow supplying jinja vars from python config.
This PR allows the user to specify a dictionary of values to `NotebookApp.jinja_template_vars` whose values are made available in the jinja templating environment.  This is primarily useful in conjunction with the `get_body_data` javascript function to allow programmatic access to Python-specified values from Javascript.  An example of where this is useful is when you want to use different constants in Javascript based on a deployment environment; this allows you to write code in your ipython_notebook_config that looks like:
```
env = os.environ['DEPLOYMENT_ENV']
if env == 'DEV':
    template_vars = dev_template_vars
...